+++ src/libanoubisprotocol/client/anoubis_client.h
@@ -47,7 +47,7 @@ void anoubis_client_close(struct anoubis_client *);
 struct anoubis_transaction * anoubis_client_close_start(
     struct anoubis_client *);
 struct anoubis_transaction * anoubis_client_policyrequest_start(
-    struct anoubis_client *, void * data, int datalen);
+    struct anoubis_client *, void * data, size_t datalen);
 int anoubis_client_process(struct anoubis_client *,
     struct anoubis_msg *);
 struct anoubis_transaction * anoubis_client_register_start(
